pyproject.toml
[project]
name = "stockcontrol-mlworker"
version = "0.1.0"
description = "ML Worker for StockControl - Background processing with Cloud Tasks"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "42N", email = "dev@42n.io" }]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.0",

    # Google Cloud
    "google-cloud-storage>=2.18.0",
    "google-auth>=2.35.0",
    "google-cloud-tasks>=2.16.0",

    # ML / Computer Vision (torch CPU installed separately in Dockerfile)
    "ultralytics>=8.3.0",
    "pillow>=10.4.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.10.0",

    # Observability
    "structlog>=24.4.0",
    "python-json-logger>=2.0.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.9",  # For file uploads in FastAPI
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
